+++ fpga/tool/vivado.py
@@ -21,7 +21,7 @@
 Implements the support of Vivado (Xilinx).
 """
 
-from fpga.tool import Tool, find_bitstream, run
+from fpga.tool import Tool, run
 
 _TEMPLATES = {
     'fpga': """\
@@ -32,7 +32,8 @@ set obj [lindex [get_hw_devices [current_hw_device]] 0]
 set_property PROGRAM.FILE #BITSTREAM# $obj
 program_hw_devices $obj
 """,
-    'detect': """open_hw
+    'detect': """\
+if { [ catch { open_hw_manager } ] } { open_hw }
 connect_hw_server
 open_hw_target
 puts [get_hw_devices]
@@ -50,13 +51,21 @@ class Vivado(Tool):
     _GEN_COMMAND = 'vivado -mode batch -notrace -quiet -source vivado.tcl'
     _TRF_COMMAND = 'vivado -mode batch -notrace -quiet -source vivado-prog.tcl'
 
+    _BIT_EXT = ['bit']
     _DEVTYPES = ['fpga', 'detect']
+
+    _GENERATED = [
+        # directories
+        '*.cache', '*.hw', '*.ip_user_files', '*.runs', '*.sim', '.Xil',
+        # files
+        '*.bit', '*.jou', '*.log', '*.rpt', 'vivado_*.zip',
+        'vivado.tcl'
+    ]
 
     def transfer(self, devtype, position, part, width, capture):
         super().transfer(devtype, position, part, width, capture)
         temp = _TEMPLATES[devtype]
         if devtype != 'detect':
-            bitstream = find_bitstream('bit')
-            temp = temp.replace('#BITSTREAM#', bitstream)
+            temp = temp.replace('#BITSTREAM#', self.bitstream)
         open("vivado-prog.tcl", 'w').write(temp)
         return run(self._TRF_COMMAND, capture)